+++ tests/04_admin.sh
@@ -9,9 +9,20 @@ cd $(dirname ${BASH_SOURCE[0]})/..
 
 ./bin/stack.sh start
 
-# Test the admin interface - use Curl
+# Wait until worker has booted
+# @todo add a time limit...
+while [ ! -f ./app/var/bootstrap_ok ]; do
+  sleep 1
+  echo .
+done
 
-# @todo ...
+# Test the admin interface - use Curl...
+
+HOST=http://localhost
+CURL=curl -v --fail
+
+# @todo test logging it to at least one db
+${CURL} ${HOST}/admin/
 
 # Stop the stack